Move updated channel balance to utils
